Break out ModelingSidebarButton to its own file
